safari-600.7-branch/LayoutTests/css3/viewport-percentage-lengths/viewport-percentage-lengths-percent-size-child-expected.txt

Test that percentage dimensions work in a child with a parent with viewport unit dimensions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS innerWidth is 800
PASS innerHeight is 600
PASS getComputedStyle(container).width is "400px"
PASS getComputedStyle(container).height is "300px"
PASS getComputedStyle(percent).width is "200px"
PASS getComputedStyle(percent).height is "150px"
PASS innerWidth is 900
PASS innerHeight is 600
PASS getComputedStyle(container).width is "450px"
PASS getComputedStyle(container).height is "300px"
PASS getComputedStyle(percent).width is "225px"
PASS getComputedStyle(percent).height is "150px"
PASS innerWidth is 900
PASS innerHeight is 640
PASS getComputedStyle(container).width is "450px"
PASS getComputedStyle(container).height is "320px"
PASS getComputedStyle(percent).width is "225px"
PASS getComputedStyle(percent).height is "160px"
PASS innerWidth is 500
PASS innerHeight is 640
PASS getComputedStyle(container).width is "250px"
PASS getComputedStyle(container).height is "320px"
PASS getComputedStyle(percent).width is "125px"
PASS getComputedStyle(percent).height is "160px"
PASS innerWidth is 800
PASS innerHeight is 600
PASS getComputedStyle(container).width is "400px"
PASS getComputedStyle(container).height is "300px"
PASS getComputedStyle(percent).width is "200px"
PASS getComputedStyle(percent).height is "150px"
